Ensure working shp ingest on EMR
