+++ empirical-data/common/src/main/scala/com/azavea/ingest/common/CSVtoSimpleFeature.scala
@@ -36,3 +36,4 @@ object CSVtoSimpleFeature {
     brMaybe.get.close
   }
 }
+
