+++ empirical-data/geomesa/src/main/scala/com/azavea/ingest/geomesa/CommandLine.scala
@@ -89,5 +89,5 @@ expression
 Note: `date' takes a format string compatible with java.text.SimpleDateFormat.
 """)
   }
-
 }
+
