+++ empirical-data/geomesa/src/main/scala/com/azavea/ingest/geomesa/Ingest.scala
@@ -83,7 +83,7 @@ object Ingest {
       var fw: FeatureWriter[SimpleFeatureType, SimpleFeature] = null
       try {
         fw = ds.getFeatureWriterAppend(params.featureName, Transaction.AUTO_COMMIT)
-        featureIter.toStream.foreach({ feature: SimpleFeature =>
+        featureIter.buffer.foreach({ feature: SimpleFeature =>
             println(feature.toString)
             val toWrite = fw.next()
             toWrite.setAttributes(feature.getAttributes)
@@ -104,3 +104,4 @@ object Ingest {
       }
     })
 }
+
