+++ empirical-data/geomesa/src/main/scala/com/azavea/ingest/geomesa/Main.scala
@@ -1,13 +1,16 @@
 package com.azavea.ingest.geomesa
 
 import org.opengis.feature.simple._
-import com.azavea.ingest.common._
 import org.geotools.feature.simple._
 import org.geotools.data.{DataStoreFinder, DataUtilities, FeatureWriter, Transaction}
 import org.apache.spark.{SparkConf, SparkContext}
 import org.apache.spark.rdd._
 
 import geotrellis.spark.util.SparkUtils
+
+import com.azavea.ingest.common._
+import com.azavea.ingest.common.csv.HydrateRDD._
+import com.azavea.ingest.common.shp.HydrateRDD._
 
 object Main {
   def main(args: Array[String]): Unit = {
@@ -26,20 +29,20 @@ object Main {
 
     params.csvOrShp match {
       case Ingest.SHP => {
-        val urls = HydrateRDD.getShpUrls(params.s3bucket, params.s3prefix)
-        val shpUrlRdd = HydrateRDD.shpUrlsToRdd(urls)
-        val shpSimpleFeatureRdd: RDD[SimpleFeature] = HydrateRDD.normalizeShpRdd(shpUrlRdd, params.featureName)
+        val urls = getShpUrls(params.s3bucket, params.s3prefix)
+        val shpUrlRdd = shpUrlsToRdd(urls)
+        val shpSimpleFeatureRdd: RDD[SimpleFeature] = NormalizeRDD.normalizeFeatureName(shpUrlRdd, params.featureName)
 
         Ingest.registerSFT(params)(shpSimpleFeatureRdd.first.getType)
         Ingest.ingestRDD(params)(shpSimpleFeatureRdd)
       }
       case Ingest.CSV => {
-        val urls = HydrateRDD.getCsvUrls(params.s3bucket, params.s3prefix, params.csvExtension)
+        val urls = getCsvUrls(params.s3bucket, params.s3prefix, params.csvExtension)
         val tybuilder = new SimpleFeatureTypeBuilder
         tybuilder.setName(params.featureName)
         params.codec.genSFT(tybuilder)
         val sft = tybuilder.buildFeatureType
-        val csvRdd: RDD[SimpleFeature] = HydrateRDD.csvUrlsToRdd(urls, params.featureName, params.codec, params.dropLines, params.separator)
+        val csvRdd: RDD[SimpleFeature] = csvUrlsToRdd(urls, params.featureName, params.codec, params.dropLines, params.separator)
 
         Ingest.registerSFT(params)(sft)
         Ingest.ingestRDD(params)(csvRdd)
@@ -47,3 +50,4 @@ object Main {
     }
   }
 }
+
